Added 64 bits support and Offsets based on unity version

+++ src/HackF5.UnitySpy/Detail/ManagedObjectInstance.cs
@@ -4,7 +4,7 @@
 
     public abstract class ManagedObjectInstance : MemoryObject, IManagedObjectInstance
     {
-        protected ManagedObjectInstance(AssemblyImage image, uint address)
+        protected ManagedObjectInstance(AssemblyImage image, IntPtr address)
             : base(image, address)
         {
         }
